parser/htmlparser/tests/mochitest/test_bug717180.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717180
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 717180</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717180">Mozilla Bug 717180</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe></iframe>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 717180 **/

SimpleTest.waitForExplicitFinish();

window.addEventListener("load", function() {
  var iframe = document.getElementsByTagName("iframe")[0];
  var d = iframe.contentDocument;
  d.open();
  iframe.addEventListener("load", function() {
    is(iframe.contentDocument.body.textContent, "SUCCESS\n", "Wrong text");
    SimpleTest.finish();
  });
  d.write("\u003Cscript>");
  d.write("window.location = 'file_bug717180.html';");
  d.write("\u003C/script>");
  d.write("FAIL");
  d.close();
  is(iframe.contentDocument.body, null, "The document should not have a body right now.");
});

</script>
</pre>
</body>
</html>
